+++ packages/core/src/components/Avatar/__stories__/Avatar.stories.tsx
@@ -11,16 +11,59 @@ import "./Avatar.stories.scss";
 
 const metaSettings = createStoryMetaSettingsDecorator({
   component: Avatar,
-  iconPropNamesArray: ["icon"]
+  iconPropNamesArray: ["icon"],
+  actionPropsArray: ["onClick"]
 });
 
 const avatarTemplate = createComponentTemplate(Avatar);
+const AvatarBadgePropsType = {
+  summary: "AvatarBadgeProps",
+  detail: `{
+    src?: string;
+    icon?: SubIcon;
+    tabIndex?: string | number;
+    className?: string;
+    size?: "xs" | "small" | "medium" | "large";
+  }`
+};
 
 export default {
   title: "Components/Avatar",
   component: Avatar,
-  argTypes: metaSettings.argTypes,
-  decorators: metaSettings.decorators
+  argTypes: {
+    ...metaSettings.argTypes,
+    bottomRightBadgeProps: {
+      control: "object",
+      table: {
+        type: AvatarBadgePropsType
+      }
+    },
+    bottomLeftBadgeProps: {
+      control: "object",
+      table: {
+        type: AvatarBadgePropsType
+      }
+    },
+    topLeftBadgeProps: {
+      control: "object",
+      table: {
+        type: AvatarBadgePropsType
+      }
+    },
+    topRightBadgeProps: {
+      control: "object",
+      table: {
+        type: AvatarBadgePropsType
+      }
+    }
+  },
+  parameters: {
+    docs: {
+      liveEdit: {
+        scope: { person1 }
+      }
+    }
+  }
 };
 
 export const Overview = {
@@ -49,9 +92,7 @@ export const Size = {
       <Avatar size="medium" src={person1} type="img" ariaLabel="Julia Martinez" />
       <Avatar size="large" src={person1} type="img" ariaLabel="Julia Martinez" />
     </Flex>
-  ),
-
-  name: "Size"
+  )
 };
 
 export const Disable = {
@@ -62,8 +103,7 @@ export const Disable = {
       <Avatar size="medium" src={person1} type="img" disabled />
       <Avatar size="large" src={person1} type="img" disabled />
     </Flex>
-  ),
-  name: "Disable"
+  )
 };
 
 export const AvatarWithText = {
@@ -74,9 +114,7 @@ export const AvatarWithText = {
       <Avatar size="medium" type="text" text="RM" backgroundColor="lipstick" ariaLabel="Ron Meir" />
       <Avatar size="large" type="text" text="RM" backgroundColor="done-green" ariaLabel="Ron Meir" />
     </Flex>
-  ),
-
-  name: "Avatar with text"
+  )
 };
 
 export const SquareAvatar = {
@@ -87,9 +125,7 @@ export const SquareAvatar = {
       <Avatar size="medium" type="icon" icon={WhatsNew} backgroundColor="aquamarine" square ariaLabel="Present" />
       <Avatar size="large" type="text" text="RM" backgroundColor="working_orange" square ariaLabel="Ron Meir" />
     </Flex>
-  ),
-
-  name: "Square avatar"
+  )
 };
 
 export const AvatarWithRightBadge = {
@@ -99,8 +135,13 @@ export const AvatarWithRightBadge = {
       <Avatar size="large" type="img" src={person1} bottomRightBadgeProps={{ src: owner }} ariaLabel="Julia Martinez" />
     </Flex>
   ),
-
-  name: "Avatar with right badge"
+  parameters: {
+    docs: {
+      liveEdit: {
+        scope: { guest, owner }
+      }
+    }
+  }
 };
 
 export const AvatarWithLeftBadge = {
@@ -110,8 +151,13 @@ export const AvatarWithLeftBadge = {
       <Avatar size="large" type="img" src={person1} bottomLeftBadgeProps={{ src: minus }} ariaLabel="Julia Martinez" />
     </Flex>
   ),
-
-  name: "Avatar with left badge"
+  parameters: {
+    docs: {
+      liveEdit: {
+        scope: { home, minus }
+      }
+    }
+  }
 };
 
 export const AvatarWithTooltip = {
@@ -145,8 +191,13 @@ export const AvatarWithTooltip = {
       </StoryDescription>
     </Flex>
   ),
-
-  name: "Avatar with tooltip"
+  parameters: {
+    docs: {
+      liveEdit: {
+        scope: { StoryDescription }
+      }
+    }
+  }
 };
 
 export const ClickableAvatar = {
@@ -165,9 +216,7 @@ export const ClickableAvatar = {
         </Flex>
       </Flex>
     );
-  },
-
-  name: "Clickable avatar"
+  }
 };
 
 export const MultipleAvatars = {
@@ -178,6 +227,11 @@ export const MultipleAvatars = {
       <Avatar type="img" src={person3} ariaLabel="Liam Caldwell" />
     </AvatarGroup>
   ),
-
-  name: "Multiple avatars"
+  parameters: {
+    docs: {
+      liveEdit: {
+        scope: { person2, person3 }
+      }
+    }
+  }
 };