+++ packages/core/src/components/Button/Button.tsx
@@ -11,7 +11,8 @@ import {
   BUTTON_ICON_SIZE,
   ButtonColor as ButtonColorEnum,
   ButtonInputType as ButtonInputTypeEnum,
-  ButtonType as ButtonTypeEnum
+  ButtonType as ButtonTypeEnum,
+  SMALL_BUTTON_ICON_SIZE
 } from "./ButtonConstants";
 import { ButtonColor, ButtonInputType, ButtonType, ButtonSize } from "./Button.types";
 import { getParentBackgroundColorNotTransparent, TRANSPARENT_COLOR } from "./helper/dom-helpers";
@@ -380,20 +381,19 @@ const Button: VibeComponent<ButtonProps, unknown> & {
       ariaPressed
     ]);
 
-    const leftIconSize = useMemo(() => {
-      if (typeof leftIcon !== "function") return;
-      return BUTTON_ICON_SIZE;
-    }, [leftIcon]);
-
-    const rightIconSize = useMemo(() => {
-      if (typeof rightIcon !== "function") return;
-      return BUTTON_ICON_SIZE;
-    }, [rightIcon]);
-
-    const successIconSize = useMemo(() => {
-      if (typeof successIcon !== "function") return;
-      return BUTTON_ICON_SIZE;
-    }, [successIcon]);
+    const iconSize = useCallback(
+      (icon: SubIcon) => {
+        if (typeof icon !== "function") return;
+        switch (size) {
+          case "xxs":
+          case "xs":
+            return SMALL_BUTTON_ICON_SIZE;
+          default:
+            return BUTTON_ICON_SIZE;
+        }
+      },
+      [size]
+    );
 
     const hasRenderableChildren = useMemo(() => React.Children.toArray(children).some(Boolean), [children]);
 
@@ -404,7 +404,7 @@ const Button: VibeComponent<ButtonProps, unknown> & {
             <Icon
               iconType="font"
               icon={leftIcon}
-              iconSize={leftIconSize}
+              iconSize={iconSize(leftIcon)}
               className={cx({
                 [styles.leftIcon]: hasRenderableChildren
               })}
@@ -416,7 +416,7 @@ const Button: VibeComponent<ButtonProps, unknown> & {
             <Icon
               iconType="font"
               icon={rightIcon}
-              iconSize={rightIconSize}
+              iconSize={iconSize(rightIcon)}
               className={cx({
                 [styles.rightIcon]: hasRenderableChildren
               })}
@@ -425,7 +425,7 @@ const Button: VibeComponent<ButtonProps, unknown> & {
           ) : null}
         </>
       ),
-      [children, hasRenderableChildren, leftIcon, leftIconSize, rightIcon, rightIconSize]
+      [children, hasRenderableChildren, iconSize, leftIcon, rightIcon]
     );
 
     if (loading) {
@@ -449,7 +449,7 @@ const Button: VibeComponent<ButtonProps, unknown> & {
               <Icon
                 iconType="font"
                 icon={successIcon}
-                iconSize={successIconSize}
+                iconSize={iconSize(successIcon)}
                 className={cx({
                   [styles.leftIcon]: !!successText
                 })}